修复潜在的数据溢出问题

+++ Demo/tinyserver/tinyserver/sockets.h
@@ -63,7 +63,7 @@ int Poll(struct pollfd *fds, unsigned long nfds, int timeout);
 
 ssize_t Read(int fd, void *buf, size_t nbytes);
 ssize_t Write(int fd, const void *buf, size_t nbytes);
-ssize_t Writen(int sockfd, const void *buf, size_t nbytes);
+int64_t Writen(int sockfd, const void *buf, int64_t nbytes);
 
 void ShutdownWrite(int sockfd);
